pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "log-analyzer-api"
version = "1.0.0"
description = "Log Analyzer API using LangGraph and AI models"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "gunicorn>=21.2.0",
    "langgraph>=0.0.32",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.5",
    "langchain-groq>=0.0.3",
    "langchain-community>=0.0.10",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.2",
    "sse-starlette>=1.8.2",
    "tavily-python>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/log-analyzer-api"
Documentation = "https://github.com/yourusername/log-analyzer-api/blob/main/README.md"
Repository = "https://github.com/yourusername/log-analyzer-api.git"
Issues = "https://github.com/yourusername/log-analyzer-api/issues"

[tool.setuptools]
packages = ["app", "app.agent", "app.api"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true